+++ src/Faithlife.Build/AppRunner.cs
@@ -1,3 +1,4 @@
+using System.Collections.Generic;
 using System.Linq;
 using McMaster.Extensions.CommandLineUtils;
 using SimpleExec;
@@ -17,7 +18,14 @@ namespace Faithlife.Build
 		/// </summary>
 		/// <param name="path">The path of the command-line app.</param>
 		/// <param name="args">The arguments to send to the command-line app.</param>
-		public static void RunApp(string path, params string[] args) =>
+		public static void RunApp(string path, params string[] args) => RunApp(path, args.AsEnumerable());
+
+		/// <summary>
+		/// Runs the specified command-line app.
+		/// </summary>
+		/// <param name="path">The path of the command-line app.</param>
+		/// <param name="args">The arguments to send to the command-line app.</param>
+		public static void RunApp(string path, IEnumerable<string> args) =>
 			Command.Run(path, ArgumentEscaper.EscapeAndConcatenate(args.Where(x => x != null)));
 
 		/// <summary>
@@ -26,7 +34,15 @@ namespace Faithlife.Build
 		/// <param name="path">The path of the command-line app.</param>
 		/// <param name="args">The arguments to send to the command-line app.</param>
 		/// <remarks>On Linux and macOS, Mono is used to run the app.</remarks>
-		public static void RunDotNetFrameworkApp(string path, params string[] args)
+		public static void RunDotNetFrameworkApp(string path, params string[] args) => RunDotNetFrameworkApp(path, args.AsEnumerable());
+
+		/// <summary>
+		/// Runs the specified .NET Framework command-line app.
+		/// </summary>
+		/// <param name="path">The path of the command-line app.</param>
+		/// <param name="args">The arguments to send to the command-line app.</param>
+		/// <remarks>On Linux and macOS, Mono is used to run the app.</remarks>
+		public static void RunDotNetFrameworkApp(string path, IEnumerable<string> args)
 		{
 			if (BuildEnvironment.IsUnix())
 			{