+++ src/Faithlife.Build/DotNetRunner.cs
@@ -1,3 +1,5 @@
+using System.Collections.Generic;
+using System.Linq;
 using McMaster.Extensions.CommandLineUtils;
 using static Faithlife.Build.AppRunner;
 
@@ -14,7 +16,13 @@ namespace Faithlife.Build
 		/// <summary>
 		/// Runs <c>dotnet</c> with the specified arguments.
 		/// </summary>
-		/// <param name="args">The arguments, if any.</param>
-		public static void RunDotNet(params string[] args) => RunApp(DotNetExe.FullPath, args);
+		/// <param name="args">The command-line arguments.</param>
+		public static void RunDotNet(params string[] args) => RunDotNet(args.AsEnumerable());
+
+		/// <summary>
+		/// Runs <c>dotnet</c> with the specified arguments.
+		/// </summary>
+		/// <param name="args">The command-line arguments.</param>
+		public static void RunDotNet(IEnumerable<string> args) => RunApp(DotNetExe.FullPath, args);
 	}
 }